requirements.txt
# Moltbook Daily Workflow Requirements

# HTTP requests
requests>=2.31.0

# Environment variables (for .env file support)
python-dotenv>=1.0.0

# Gradio for web UI (HF Spaces app)
gradio>=4.0.0

# Note: LLM providers are accessed via OpenRouter API
# No direct openai or anthropic packages needed
# Only OPENROUTER_API_KEY is required for LLM functionality
